requirements.txt
asgiref==3.6.0
Django==4.1.7
django-jet-reboot==1.3.3
feedparser==6.0.10
mysqlclient==2.1.1
Pillow==9.4.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
